requirements.txt
Flask
flask_cors
flake8
python-dotenv
pandas
scikit-learn
beautifulsoup4
